requirements.txt
numpy==1.18.1
scipy==1.4.1
scikit_learn==0.22.2
Pillow==7.1.2
wordcloud==1.7.0
jieba==0.42.1